Add files via upload
The purpose of this project is to develop a computer vision model using deep learning 
techniques that can detect objects in a live video and when given image or video as input. 
Recognizing objects in a video is a difficult task and the difficulty increases if the 
inference is expected in real-time. The project begins with analyzing the advantages and 
disadvantages of existing object detection architectures on the self-prepared dataset. The 
goal is to come up with an optimized object detection architecture using the advantages of 
existing architectures. The developed architecture may be used for various applications in 
real-time with better accuracy and lesser time is taken for generating the inference. 
The superiority of the architecture will be demonstrated by developing a computer vision 
model that recognizes objects belonging to different classes(College buildings in our 
case). The model will be deployed to detect objects in a live streaming video or with an 
image or video as the input.